+++ src/aui-form-builder-II/tests/unit/js/tests.js
@@ -29,22 +29,22 @@ YUI.add('aui-form-builder-II-tests', function(Y) {
         'should add a new field type on form': function() {
             var formBuilder = this._formBuilder;
 
-            formBuilder.registerFieldType([{
+            formBuilder.registerFieldTypes([{
                 icon: 'icon1'
             }]);
             Y.Assert.isNotNull(Y.one('.icon1'));
 
-            formBuilder.registerFieldType({icon: 'icon2'});
+            formBuilder.registerFieldTypes({icon: 'icon2'});
             Y.Assert.isNotNull(Y.one('.icon2'));
 
-            formBuilder.registerFieldType([
+            formBuilder.registerFieldTypes([
                 new Y.FormBuilderIIFieldType({
                     icon: 'icon3'
                 })
             ]);
             Y.Assert.isNotNull(Y.one('.icon3'));
 
-            formBuilder.registerFieldType(
+            formBuilder.registerFieldTypes(
                 new Y.FormBuilderIIFieldType({
                     icon: 'icon4'
                 })
@@ -58,18 +58,60 @@ YUI.add('aui-form-builder-II-tests', function(Y) {
                     icon: 'icon-test'
                 });
 
-            formBuilder.registerFieldType(fieldType1);
+            formBuilder.registerFieldTypes(fieldType1);
 
-
-            formBuilder.unregisterFieldType('icon-test');
-            Y.Assert.isNotNull(Y.one('.icon-test'));
-
-            formBuilder.unregisterFieldType(new Y.FormBuilderIIFieldType({icon: 'icon-test'}));
+            formBuilder.unregisterFieldTypes(new Y.FormBuilderIIFieldType({icon: 'icon-test'}));
             Y.Assert.isNotNull(Y.one('.icon-test'));
 
             Y.Assert.isNotNull(Y.one('.icon-test'));
-            formBuilder.unregisterFieldType(fieldType1);
+            formBuilder.unregisterFieldTypes(fieldType1);
             Y.Assert.isNull(Y.one('.icon-test'));
+        },
+
+        'should remove field types by field class': function() {
+            this._formBuilder.registerFieldTypes([
+                {
+                    fieldClass: 'Class1'
+                },
+                {
+                    fieldClass: 'Class2'
+                },
+                {
+                    fieldClass: 'Class1'
+                }
+            ]);
+
+            this._formBuilder.unregisterFieldTypes('Class1');
+            Y.Assert.areEqual(1, this._formBuilder.get('fieldTypes').length);
+
+            this._formBuilder.unregisterFieldTypes('Class3');
+            Y.Assert.areEqual(1, this._formBuilder.get('fieldTypes').length);
+        },
+
+        'should remove multiple field types on same call': function() {
+            var fieldType1 = new Y.FormBuilderIIFieldType({
+                fieldClass: 'Class1'
+            });
+
+            this._formBuilder.registerFieldTypes([
+                fieldType1,
+                {
+                    fieldClass: 'Class2'
+                },
+                {
+                    fieldClass: 'Class2'
+                },
+                {
+                    fieldClass: 'Class3'
+                }
+            ]);
+
+            this._formBuilder.unregisterFieldTypes(['Class2', fieldType1]);
+            Y.Assert.areEqual(1, this._formBuilder.get('fieldTypes').length);
+            Y.Assert.areEqual(
+                'Class3',
+                this._formBuilder.get('fieldTypes')[0].get('fieldClass')
+            );
         },
 
         'should show the form builder modal': function() {